requirements.txt
streamlit
upstash-redis
langchain
langchain-community
langchain-together
faiss-cpu
huggingface-hub
python-docx
PyPDF2
pillow
pytesseract
numpy
tensorflow
keras
absl-py==2.1.0
aiodns==3.0.0
aiofiles==23.1.0
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audio-metadata==0.12.2
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==23.12.1
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
brotli==1.1.0
bs4==0.0.2
CacheControl==0.13.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
contourpy==1.2.0
cookiecutter==2.5.0
cryptography==41.0.7
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.8
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-cli==0.18.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dependency-injector==4.41.0
dill==0.3.7
dispatcher==0.3.0
distlib==0.3.8
distro==1.8.0
dnspython==2.4.2
docker==7.0.0
dotenv==0.0.5
duro==0.2.2
email-validator==2.1.0.post1
environs==11.0.0
epiparser==1.5.2
epkg==0.0.1
epydoc==3.0.1
eventlet==0.35.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flask==3.0.0
flask-cors==4.0.0
flask-restful==0.3.10
flask-wtf==1.2.1
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.0
fsspec==2023.10.0
furl==2.1.3
gast==0.5.4
gensim==4.3.2
gitdb==4.0.11
gitpython==3.1.40
graphviz==0.20.1
greenlet==3.0.3
griffin==0.1.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.1
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
importlib-metadata==6.11.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipython==8.18.1
ipywidgets==8.1.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
jieba==0.42.1
jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-client==8.6.0
jupyter-core==5.7.0
jupyterlab-pygments==0.3.0
jupyterlab-widgets==3.0.9
keras==3.0.5
kiwisolver==1.4.5
lazy-loader==0.3
lightgbm==4.1.0
linkify-it-py==2.0.2
lit==17.0.5
llvmlite==0.41.1
lockfile==0.12.2
lxml==5.0.1
lz4==4.3.2
markdown==3.5.1
markupsafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mlflow==2.9.2
msgpack==1.0.7
murmurhash==1.0.10
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook-shim==0.2.3
numba==0.58.1
numpy==1.26.2
openai==0.28.1
opencv-python==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.10
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pip==23.3.2
protobuf==4.25.1
psutil==5.9.6
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pyparsing==3.1.1
pysocks==1.7.1
python-dateutil==2.8.2
pyyaml==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.15.2
spacy==3.7.2
tensorflow==2.15.0
torch==2.1.2
transformers==4.36.2
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
